+++ lib/react_on_rails/configuration.rb
@@ -1,17 +1,51 @@
 module ReactOnRails
   def self.configure
     yield(configuration)
+    setup_config_values
   end
 
-  DEFAULT_GENERATED_ASSETS_DIRS = [
-    %w(app assets javascripts generated),
-    %w(app assets javascripts stylesheets generated)
-  ].map { |dirs| File.join(*dirs) }.freeze
+  DEFAULT_GENERATED_ASSETS_DIR = File.join(%w(app assets webpack)).freeze
+
+  def self.setup_config_values
+    if @configuration.webpack_generated_files.empty?
+      files = ["client-bundle.js"]
+      if @configuration.server_bundle_js_file.present?
+        files << @configuration.server_bundle_js_file
+      end
+      @configuration.webpack_generated_files = files
+    end
+
+    if @configuration.generated_assets_dirs.present?
+      puts "[DEPRECATION] ReactOnRails: Use config.generated_assets_dir rather than "\
+        "generated_assets_dirs"
+      if @configuration.generated_assets_dir.blank?
+        @configuration.generated_assets_dir = @configuration.generated_assets_dirs
+      else
+        puts "[DEPRECATION] ReactOnRails. You have both generated_assets_dirs and "\
+          "generated_assets_dir defined. Define ONLY generated_assets_dir"
+      end
+    end
+
+    if @configuration.generated_assets_dir.blank?
+      @configuration.generated_assets_dir = DEFAULT_GENERATED_ASSETS_DIR
+      puts "ReactOnRails: Set generated_assets_dir to default: #{DEFAULT_GENERATED_ASSETS_DIR}"
+    end
+
+    if @configuration.server_bundle_js_file.include?(File::SEPARATOR)
+      puts "[DEPRECATION] ReactOnRails: remove path from server_bundle_js_file in configuration. "\
+        "All generated files must go in #{@configuration.generated_assets_dir}"
+      @configuration.server_bundle_js_file = File.basename(@configuration.server_bundle_js_file)
+    end
+  end
 
   def self.configuration
     @configuration ||= Configuration.new(
-      generated_assets_dirs: DEFAULT_GENERATED_ASSETS_DIRS,
-      server_bundle_js_file: File.join(*%w(app assets javascripts generated server.js)),
+      generated_assets_dirs: nil,
+
+      # generated_assets_dirs is deprecated
+      generated_assets_dir: "",
+
+      server_bundle_js_file: "",
       prerender: false,
       replay_console: true,
       logging_on_server: true,
@@ -20,7 +54,9 @@ module ReactOnRails
       development_mode: Rails.env.development?,
       server_renderer_pool_size: 1,
       server_renderer_timeout: 20,
-      skip_display_none: false)
+      skip_display_none: false,
+      webpack_generated_files: []
+    )
   end
 
   class Configuration
@@ -28,18 +64,19 @@ module ReactOnRails
                   :trace, :development_mode,
                   :logging_on_server, :server_renderer_pool_size,
                   :server_renderer_timeout, :raise_on_prerender_error,
-                  :skip_display_none, :generated_assets_dirs
+                  :skip_display_none, :generated_assets_dirs, :generated_assets_dir,
+                  :webpack_generated_files
 
     def initialize(server_bundle_js_file: nil, prerender: nil, replay_console: nil,
                    trace: nil, development_mode: nil,
                    logging_on_server: nil, server_renderer_pool_size: nil,
                    server_renderer_timeout: nil, raise_on_prerender_error: nil,
-                   skip_display_none: nil, generated_assets_dirs: DEFAULT_GENERATED_ASSETS_DIRS)
-      self.server_bundle_js_file = if File.exist?(server_bundle_js_file)
-                                     server_bundle_js_file
-                                   end
-
+                   skip_display_none: nil, generated_assets_dirs: nil,
+                   generated_assets_dir: nil, webpack_generated_files: nil)
+      self.server_bundle_js_file = server_bundle_js_file
       self.generated_assets_dirs = generated_assets_dirs
+      self.generated_assets_dir = generated_assets_dir
+
       self.prerender = prerender
       self.replay_console = replay_console
       self.logging_on_server = logging_on_server
@@ -55,6 +92,8 @@ module ReactOnRails
       # Server rendering:
       self.server_renderer_pool_size = self.development_mode ? 1 : server_renderer_pool_size
       self.server_renderer_timeout = server_renderer_timeout # seconds
+
+      self.webpack_generated_files = webpack_generated_files
     end
   end
 end