+++ lib/react_on_rails/test_helper/webpack_assets_compiler.rb
@@ -3,30 +3,38 @@
 module ReactOnRails
   module TestHelper
     class WebpackAssetsCompiler
-      def compile
-        compile_type(:client)
-        compile_type(:server) if Utils.server_rendering_is_enabled?
+      def compile_as_necessary(stale_files)
+        compile_client(stale_files)
+        compile_server(stale_files)
+      end
+
+      def compile_client(stale_files)
+        compile_type(:client) if needs_client_compile?(stale_files)
+      end
+
+      def compile_server(stale_files)
+        compile_type(:server) if needs_server_compile?(stale_files)
       end
 
       private
 
       def compile_type(type)
-        unless @printed_msg
-          puts <<-MSG
-If you are frequently running tests, you can run webpack in watch mode to speed up this process.
-See the official documentation:
-https://github.com/shakacode/react_on_rails/blob/master/docs/additional_reading/rspec_configuration.md
-          MSG
-          @printed_msg = true
-        end
-
         puts "\nBuilding Webpack #{type}-rendering assets..."
 
         build_output = `cd client && npm run build:#{type}`
 
-        fail "Error in building assets!\n#{build_output}" unless Utils.last_process_completed_successfully?
+        raise "Error in building assets!\n#{build_output}" unless Utils.last_process_completed_successfully?
 
         puts "Completed building Webpack #{type}-rendering assets."
+      end
+
+      def needs_client_compile?(stale_files)
+        !stale_files.all? { |name| name.include?("server") }
+      end
+
+      def needs_server_compile?(stale_files)
+        return false unless Utils.server_rendering_is_enabled?
+        stale_files.any? { |name| name.include?("server") }
       end
     end
   end