+++ spec/dummy/app/assets/javascripts/application_with_turbolinks.js
@@ -1,7 +1,0 @@
-// This file is so that we can test the dummy app with turbolinks both enabled
-// and disabled via an environment variable (see also:
-//           Gemfile, layout/application.html.erb, and config/initializers/assets.rb
-
-//= require application
-
-//= require turbolinks