+++ spec/dummy/config/initializers/assets.rb
@@ -10,6 +10,16 @@ Rails.application.config.assets.version = "1.0"
 # application.js, application.css, and all non-JS/CSS in app/assets folder are already added.
 # Rails.application.config.assets.precompile += %w( search.js )
 
-unless ENV["DISABLE_TURBOLINKS"].present?
-  Rails.application.config.assets.precompile += %w( application_with_turbolinks.js )
-end
+# Add folder with webpack generated assets to assets.paths
+Rails.application.config.assets.paths << Rails.root.join("app", "assets", "webpack")
+
+# Precompile additional assets.
+# application.js, application.css, and all non-JS/CSS in app/assets folder are already added.
+Rails.application.config.assets.precompile << "server-bundle.js"
+
+type = ENV["REACT_ON_RAILS_ENV"] == "HOT" ? "non_webpack" : "static"
+Rails.application.config.assets.precompile +=
+  [
+    "application_#{type}.js",
+    "application_#{type}.css"
+  ]