+++ spec/dummy/spec/requests/server_render_check_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-describe "Server Rendering" do
+describe "Server Rendering", :server_rendering do
   it "generates server rendered HTML if server renderering enabled" do
     get server_side_hello_world_with_options_path
     html_nodes = Nokogiri::HTML(response.body)
@@ -29,7 +29,7 @@ describe "Server Rendering" do
   end
 
   describe "reloading the server bundle" do
-    let(:server_bundle) { File.expand_path("../../../app/assets/javascripts/generated/server.js", __FILE__) }
+    let(:server_bundle) { File.expand_path("../../../app/assets/webpack/server-bundle.js", __FILE__) }
     let!(:original_bundle_text) { File.read(server_bundle) }
     before do
       ReactOnRails.configure { |config| config.development_mode = true }