+++ spec/react_on_rails/generators/install_generator_spec.rb
@@ -221,7 +221,7 @@ describe InstallGenerator, type: :generator do
         node_modules
 
         # Generated js bundles
-        /app/assets/javascripts/generated/*
+        /app/assets/webpack/*
 
         MSG
       expect(GeneratorMessages.output)