+++ spec/react_on_rails/support/shared_examples/base_generator_examples.rb
@@ -22,7 +22,7 @@ shared_examples "base_generator:base" do |options|
       node_modules
 
       # Generated js bundles
-      /app/assets/javascripts/generated/*
+      /app/assets/webpack/*
     MATCH
     assert_file ".gitignore", match
   end
@@ -32,11 +32,11 @@ shared_examples "base_generator:base" do |options|
       // DO NOT REQUIRE jQuery or jQuery-ujs in this file!
       // DO NOT REQUIRE TREE!
 
-      // CRITICAL that generated/vendor-bundle must be BEFORE bootstrap-sprockets and turbolinks
+      // CRITICAL that vendor-bundle must be BEFORE bootstrap-sprockets and turbolinks
       // since it is exposing jQuery and jQuery-ujs
 
-      //= require generated/vendor-bundle
-      //= require generated/app-bundle
+      //= require vendor-bundle
+      //= require app-bundle
 
     MATCH
     assert_file("app/assets/javascripts/application.js") do |contents|
@@ -87,16 +87,7 @@ shared_examples "base_generator:base" do |options|
   end
 
   it "appends path configurations to assets.rb" do
-    expected = <<-EXPECTED.strip_heredoc
-      # Add client/assets/ folders to asset pipeline's search path.
-      # If you do not want to move existing images and fonts from your Rails app
-      # you could also consider creating symlinks there that point to the original
-      # rails directories. In that case, you would not add these paths here.
-      Rails.application.config.assets.paths << Rails.root.join("client", "assets", "stylesheets")
-      Rails.application.config.assets.paths << Rails.root.join("client", "assets", "images")
-      Rails.application.config.assets.paths << Rails.root.join("client", "assets", "fonts")
-      Rails.application.config.assets.precompile += %w( generated/server-bundle.js )
-    EXPECTED
+    expected = ReactOnRails::Generators::BaseGenerator::ASSETS_RB_APPEND
     assert_file("config/initializers/assets.rb") { |contents| assert_match(expected, contents) }
   end
 end
@@ -185,7 +176,7 @@ shared_examples "base_generator:server_rendering" do
   end
 
   it "sets server bundle js file to server-bundle in react_on_rails initializer" do
-    regexp = %r{config.server_bundle_js_file = "app/assets/javascripts/generated/server-bundle.js"}
+    regexp = /config.server_bundle_js_file = "server-bundle.js"/
     assert_file("config/initializers/react_on_rails.rb") do |contents|
       assert_match(regexp, contents)
     end